update typescript base path
